backend/django/orders/requirements.txt
asgiref==3.9.1
coverage==7.11.0
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
iniconfig==2.3.0
kafka-python==2.2.15
model-bakery==1.20.5
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.2
pytest==8.4.2
pytest-cov==7.0.0
pytest-django==4.11.1
sqlparse==0.5.3
